backend/requirements.txt
asgiref==3.11.0
Django==5.2.9
django-tinymce==5.0.0
pillow==12.1.0
sqlparse==0.5.5
tzdata==2025.3
